finished achievements and settings
Achievements were relatively straight forward once I ignored the mediocre advice of internet strangers. I just did it how I know how to do it, and it works just how I want it to. The one achievement I am not entirely happy with is the koffing building ach, might change how that checkfunction works at some point in future.

Working on the settings menu made me realize how much easier this could be to do in React (not to mention probably prettier). My goal with this project was to do it in just JS/HTML/CSS though and I'm sticking to that for now. I am doing this more for learning and understanding JS than I am for making a perfect website anyway. Anyway, settings menu is functional and has some localStorage settings. Good for now.